fix: filesize should be Long (files larger 2.1 GB)

+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/util/FileDescriptor.kt
@@ -3,9 +3,9 @@ package de.connect2x.trixnity.messenger.util
 import io.ktor.http.*
 import net.folivo.trixnity.utils.ByteArrayFlow
 
-interface FileDescriptor{
+interface FileDescriptor {
     val fileName: String
-    val fileSize: Int?
+    val fileSize: Long?
     val mimeType: ContentType?
     val content: ByteArrayFlow
 }
